+++ BK/src/main/java/com/searshc/mygarage/orchestrators/DashboardOrchestrator.java
@@ -37,7 +37,7 @@ public class DashboardOrchestrator extends BaseOrchestrator{
     	
     	List<FamilyVehicle> familyVehicles = this.familyVehicleService.getConfirmedFamilyVehiclesByUserId(userId);
     	
-    	double highestMileage, localMileage, NcdbMileage;
+    	int highestMileage, localMileage, NcdbMileage;
     	for(FamilyVehicle familyVehicle : familyVehicles) {
     		highestMileage = familyVehicle.getMileage();
     		localMileage = this.recordService.getHighestMileageByFamilyVehicleId(familyVehicle.getId());
